requirements.txt
aiohttp==3.9.1
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
